+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 ### Fixed
 - Storage write batches now correctly abort when any query in the batch fails.
+- Rank cache correctly calculates record expiry times.
 
 ## [2.4.2] - 2019-03-25
 ### Added